form for creating an entry and its functionality is now in its own module
