+++ src/components/library/RudimentDetails.js
@@ -1,19 +1,17 @@
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { RudimentsData } from "../data_management/RudimentsData.js";
-import { EntriesData } from "../data_management/EntriesData.js";
 import { UsersData } from "../data_management/UsersData.js";
 import { StudentsProfileData } from "../data_management/StudentsProfileData.js";
 import { Leaderboard } from "./Leaderboard.js";
 import { currentUserId } from "../data_management/Fetch.js";
 import Metronome from "@kevinorriss/react-metronome";
 import "./RudimentDetails.css";
+import { EntryForm } from "./EntryForm.js";
 
 export const RudimentDetails = () => {
     const [rudiment, setRudiment] = useState({});
     const [students, setStudents] = useState([]);
-    const [bpm, setBPM] = useState(0);
-    const [selectedStudent, setSelectedStudent] = useState(0);
     const [entrySubmitted, setSubmitState] = useState(false);
     const [isSubmitterStudent, setSubmitter] = useState(false);
     const [lbAccess, setLb] = useState(false);
@@ -41,19 +39,6 @@ export const RudimentDetails = () => {
         }
     }, [isSubmitterStudent]);
 
-    const submitEntry = () => {
-        const userId = isSubmitterStudent ? parseInt(localStorage.getItem("rude_user")) : selectedStudent;
-        const entry = {
-            bpm,
-            userId,
-            rudimentId: parseInt(rudimentId),
-            approved: !isSubmitterStudent,
-            timestamp: Date.now(),
-        };
-
-        EntriesData.postEntry(entry).then(() => setSubmitState(true));
-    };
-
     return (
         <>
             <h1>
@@ -76,29 +61,12 @@ export const RudimentDetails = () => {
                 </button>
             )}
             {!entrySubmitted && userView === "entry" ? (
-                <form>
-                    {!isSubmitterStudent && (
-                        <>
-                            <label htmlFor="studentSelector">Student</label>
-                            <select
-                                name="studentSelector"
-                                onChange={(e) => setSelectedStudent(parseInt(e.target.value))}
-                            >
-                                <option value={0}>Select A Student</option>
-                                {students.map((s) => (
-                                    <option key={s.id} value={s.userId}>
-                                        {s.user.name}
-                                    </option>
-                                ))}
-                            </select>
-                        </>
-                    )}
-                    <label htmlFor="bpm">BPM</label>
-                    <input type="number" placeholder="BPM" onChange={(e) => setBPM(+e.target.value)} />
-                    <button type="button" onClick={submitEntry}>
-                        Submit Entry
-                    </button>
-                </form>
+                <EntryForm
+                    students={students}
+                    isStudent={isSubmitterStudent}
+                    rudimentId={rudimentId}
+                    setSubmitState={setSubmitState}
+                />
             ) : entrySubmitted ? (
                 <p>Submission Complete</p>
             ) : (
